renamed stream2 to stream!
