+++ stream/README.md
@@ -1,0 +1,11 @@
+# radio-stream
+
+
+## basic usage
+
+throw some mp3s in the `assets/audio` directory then:
+
+```
+npm install
+node index.js
+```
