--- stream2/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "radio-stream",
-  "version": "1.0.0",
-  "description": "self-hosted rtmp music stream",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "dmamills",
-  "license": "MIT",
-  "dependencies": {
-    "chalk": "^2.4.2",
-    "chalkline": "0.0.5",
-    "cli-progress": "^2.1.1",
-    "fluent-ffmpeg": "^2.1.2",
-    "knex": "^0.16.3",
-    "moment": "^2.24.0",
-    "music-metadata": "^3.5.3",
-    "sqlite3": "^4.0.6"
-  }
-}
